Refactored Bonito scrapper to have the same methods as other scrappers. parseSingleGrid() is now in abstract base class.
