+++ crawler/src/test/java/com/javaacademy/crawler/jsoup/BonitoScrapperTests.java
@@ -4,6 +4,8 @@ import com.javaacademy.crawler.common.model.BookModel;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
+import java.util.Collections;
+
 /**
  * @author devas
  * @since 05.09.17
@@ -12,16 +14,24 @@ import org.testng.annotations.Test;
 public class BonitoScrapperTests {
 
     private static final String LINK = "https://bonito.pl/k-167304-podreczny-slownik-niemiecko-polski-a-z";
-
-    @Test(expectedExceptions = IllegalArgumentException.class)
-    public void testInvalidUrlScrapping() {
-        BonitoScrapper bonitoScrapper = new BonitoScrapper("");
-        Assert.assertNotNull(bonitoScrapper.scrape());
-    }
+    private static final String TITLE = "Podręczny słownik niemiecko-polski A-Z";
 
     public void shouldScrapeTitle() {
         BonitoScrapper bonitoScrapper = new BonitoScrapper();
         bonitoScrapper.connect(LINK);
-        Assert.assertEquals(bonitoScrapper.parseSinglePage(LINK).getTitle(), "Podręczny słownik niemiecko-polski A-Z");
+        Assert.assertEquals(bonitoScrapper.parseSinglePage(LINK).getTitle(), TITLE);
+    }
+
+    public void shouldScrapeNothing() {
+        BonitoScrapper bonitoScrapper = new BonitoScrapper();
+        bonitoScrapper.setPageEndIndex(0);
+        Assert.assertEquals(bonitoScrapper.scrape(), Collections.emptySet());
+    }
+
+    public void shouldScrapeSomePages() {
+        BonitoScrapper bonitoScrapper = new BonitoScrapper();
+        bonitoScrapper.setPageEndIndex(1);
+        bonitoScrapper.setShouldDataBeScrapped(false);
+        Assert.assertNotNull(bonitoScrapper.scrape());
     }
 }
